feat: moved experimental signup with api integration
