+++ src/Interfaces/newUserInterface.ts
@@ -1,0 +1,7 @@
+export interface NewUser {
+  username: string;
+  password: string;
+  name: string;
+  email: string;
+  memberShipType: string;
+}
